+++ gmime/gmime-stream-mem.c
@@ -139,7 +139,7 @@ stream_read (GMimeStream *stream, char *buf, size_t len)
 	
 	bound_end = stream->bound_end != -1 ? stream->bound_end : (gint64) mem->buffer->len;
 	
-	n = MIN (bound_end - stream->position, (gint64) len);
+	n = (size_t) MIN (bound_end - stream->position, (gint64) len);
 	if (n > 0) {
 		memcpy (buf, mem->buffer->data + stream->position, n);
 		stream->position += n;
@@ -169,7 +169,7 @@ stream_write (GMimeStream *stream, const char *buf, size_t len)
 	} else
 		bound_end = stream->bound_end;
 	
-	n = MIN (bound_end - stream->position, (gint64) len);
+	n = (size_t) MIN (bound_end - stream->position, (gint64) len);
 	if (n > 0) {
 		memcpy (mem->buffer->data + stream->position, buf, n);
 		stream->position += n;
